Rename addon to geo-routing and key the default map by region code (NA/CA/EU/AU)
